refactor(itun): decompose PlayerPilotDisplay, add aggregateActions
Story 5.4: Split PlayerPilotDisplay (757 -> 104 lines) into a thin
facade that delegates to PilotListingCard (85 lines) and PilotSheet
(699 lines). Extract ChassisBadge as shared sub-component.

Story 5.3: Add aggregateActions pure utility with 17 tests. Fix
remaining entityHelpers imports missed in prior commit.

Closes #121 #124
