+++ apps/in-the-union-now/src/hooks/usePilotSheet.ts
@@ -27,7 +27,7 @@ import { computeCrawlerStatsFromTechLevel } from '../lib/crawlerUtils'
 import { useSaveStatus } from './useSaveStatus'
 import { useRealtimeSubscription } from './useRealtimeSubscription'
 import { getEntityAccess } from '../lib/entityAccess'
-import { findChassisById } from '../lib/entityHelpers'
+import { findChassisById } from 'salvageunion-reference'
 import { extractComrades } from '../lib/comradeUtils'
 import { getErrorMessage } from '../lib/errors'
 import { changeLogApi } from '../lib/api/changeLogApi'
